Refactor subscribers service creation and move deleteSubscriber extension
- Rename getSubscribersService() to maketSubscribersService() to better reflect factory pattern
- Move deleteSubscriber extension from SubsriberDetailsViewModel to dedicated SubscribersServiceRemote+Extensions file
- Update all callers to use the renamed method
