+++ WordPress/Classes/ViewRelated/Blog/Subscribers/Details/SubsriberDetailsViewModel.swift
@@ -32,17 +32,17 @@ struct SubsriberDetailsViewModel {
     }
 
     func getDetails() async throws -> SubscribersServiceRemote.GetSubscriberDetailsResponse {
-        try await blog.getSubscribersService()
+        try await blog.maketSubscribersService()
             .getSubsciberDetails(siteID: blog.dotComSiteID, subscriberID: subscriberID)
     }
 
     func getStats() async throws -> SubscribersServiceRemote.GetSubscriberStatsResponse {
-        try await blog.getSubscribersService()
+        try await blog.maketSubscribersService()
             .getSubsciberStats(siteID: blog.dotComSiteID, subscriberID: subscriberID)
     }
 
     func delete(_ subscriber: SubscribersServiceRemote.SubsciberBasicInfoResponse) async throws {
-        try await blog.getSubscribersService()
+        try await blog.maketSubscribersService()
             .deleteSubscriber(subscriber, siteID: blog.dotComSiteID)
     }
 
@@ -50,34 +50,5 @@ struct SubsriberDetailsViewModel {
         let absolute = date.formatted(date: .abbreviated, time: .shortened)
         let relative = relativeFormatter.localizedString(for: date, relativeTo: .now)
         return absolute + " (\(relative))"
-    }
-}
-
-extension SubscribersServiceRemote {
-    static let subscriberIDKey = "subscriberIDKey"
-
-    @MainActor
-    func deleteSubscriber(_ subscriber: SubscribersServiceRemote.SubsciberBasicInfoResponse, siteID: Int) async throws {
-        let service = PeopleServiceRemote(wordPressComRestApi: wordPressComRestApi)
-        try await withUnsafeThrowingContinuation { continuation in
-            if subscriber.isDotComUser {
-                service.deleteFollower(siteID, userID: subscriber.dotComUserID, success: {
-                    continuation.resume()
-                }, failure: {
-                    continuation.resume(throwing: $0)
-                })
-            } else {
-                service.deleteEmailFollower(siteID, userID: subscriber.subscriberID, success: {
-                    continuation.resume()
-                }, failure: {
-                    continuation.resume(throwing: $0)
-                })
-            }
-        }
-        NotificationCenter.default.post(
-            name: .subscriberDeleted,
-            object: nil,
-            userInfo: [SubscribersServiceRemote.subscriberIDKey: subscriber.subscriberID]
-        )
     }
 }
