+++ WordPress/Classes/ViewRelated/Blog/Subscribers/Helpers/SubscribersBlog.swift
@@ -25,7 +25,7 @@ struct SubscribersBlog {
         SubscribersBlog(dotComSiteID: 1, getRestAPI: { nil })
     }
 
-    func getSubscribersService() throws -> SubscribersServiceRemote {
+    func maketSubscribersService() throws -> SubscribersServiceRemote {
         guard let api = getRestAPI() else {
             throw URLError(.unknown, userInfo: [NSLocalizedDescriptionKey: SharedStrings.Error.generic])
         }
